Remove the token type
The ruby library makes default calls to the API to retrieve tokens,
meaning that the most important thing for the rails client to store is
the facade, so that it retrieves the correct tokens. Some methods set
the facade to POS, so we pass the merchant facade in as an argument. The
default facade for the rails app will be the merchant facade.